6_agentcore/requirements.txt
bedrock-agentcore
strands-agents[a2a]
mcp-proxy-for-aws
httpx
requests-aws4auth
